requirements.txt
websockets==10.3
RPi.GPIO==0.7.1
